fixing unpredicted errors

+++ project/app/models/tables/Course.scala
@@ -2,7 +2,7 @@ package models.tables
 
 import play.api.libs.json._
 
-case class Course(id: Int, title: String, description: String, chapters: JsArray) {
+case class Course(id: Long, title: String, description: String, chapters: JsArray) {
   def toJson: JsObject = Json.obj(
     "id" -> id,
     "title" -> title,
@@ -13,7 +13,7 @@ case class Course(id: Int, title: String, description: String, chapters: JsArray
 
 object Course {
   implicit def fromJson(jsObject: JsValue): Course = Course(
-    (jsObject \ "id").as[Int],
+    (jsObject \ "id").as[Long],
     (jsObject \ "title").as[String],
     (jsObject \ "description").as[String],
     (jsObject \ "chapters").as[JsArray]
