+++ project/app/models/tests/Meta.scala
@@ -1,3 +1,3 @@
 package models.tests
 
-case class Meta(id: Int, title: String, description: String) extends FileContent
+case class Meta(id: Long, title: String, description: String) extends FileContent
